+++ src/types/types.ts
@@ -10,3 +10,13 @@ export type UserData = {
 	bio?: string;
 	ensName?: string;
 }
+
+export type Metadata = {
+  name?: string;
+  description?: string;
+  image?: string;
+  image_details?: any;
+  image_url?: string;
+  created_by?: string;
+  attributes?: Array<any>;
+};
